add ABC/215/b.cpp

+++ ABC/215/b.cpp
@@ -4,5 +4,5 @@ using namespace std;
 
 int main() {
     long long int N; cin >> N;
-    cout << floor(log2(N)) << endl;
+    cout << floor(log2((long double)N)) << endl;
 }
